+++ politeiawww/legacy/dcc.go
@@ -20,12 +20,12 @@ import (
 
 	"github.com/decred/politeia/cmsplugin"
 	"github.com/decred/politeia/decredplugin"
-	"github.com/decred/politeia/mdstream"
 	pd "github.com/decred/politeia/politeiad/api/v1"
 	"github.com/decred/politeia/politeiad/api/v1/identity"
 	cms "github.com/decred/politeia/politeiawww/api/cms/v1"
 	www "github.com/decred/politeia/politeiawww/api/www/v1"
 	"github.com/decred/politeia/politeiawww/legacy/cmsdatabase"
+	"github.com/decred/politeia/politeiawww/legacy/mdstream"
 	"github.com/decred/politeia/politeiawww/legacy/user"
 	"github.com/decred/politeia/util"
 )
